pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "strava-mcp"
version = "0.1.0"
description = "MCP server for Strava API integration"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "you@example.com" }
]
keywords = ["mcp", "strava", "claude", "api"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mcp>=1.0.0",
    "stravalib>=2.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "jinja2>=3.0.0",
    "python-dotenv>=1.0.0",
    "keyring>=24.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
    "respx>=0.21.0",
    "ruff>=0.1.0",
    "ty>=0.0.1a4",
]

[project.scripts]
strava-mcp = "strava_mcp.server:main"

[project.urls]
Homepage = "https://github.com/bronteee/strava-mcp"
Repository = "https://github.com/bronteee/strava-mcp"
Issues = "https://github.com/bronteee/strava-mcp/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/strava_mcp"]

[tool.hatch.build.targets.sdist]
include = [
    "/src",
    "/tests",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["src"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]
